Refactor checkboxes
